requirements.txt
python-dotenv==1.0.1
boto3==1.34.105
botocore==1.34.105
psycopg2-binary==2.9.9
pymysql==1.1.0
pandas==2.2.2
sqlalchemy==1.4.52
apache-airflow==2.9.1
pyarrow==16.1.0